SMARTCARD-170 Feature for add new Smart Card

See merge request !437

+++ app/src/main/java/com/worldventures/dreamtrips/wallet/service/command/reset/RemoveSmartCardDataCommand.java
@@ -21,6 +21,12 @@ public class RemoveSmartCardDataCommand extends Command<Void> implements Injecta
    @Inject LostCardRepository lostCardRepository;
    @Inject SmartCardInteractor smartCardInteractor;
 
+   private final boolean withPaymentCards;
+
+   public RemoveSmartCardDataCommand(boolean withPaymentCards) {
+      this.withPaymentCards = withPaymentCards;
+   }
+
    @Override
    protected void run(CommandCallback<Void> callback) throws Throwable {
       smartCardInteractor.activeSmartCardPipe()
@@ -32,8 +38,10 @@ public class RemoveSmartCardDataCommand extends Command<Void> implements Injecta
    private Observable<Void> removeCache() {
       return Observable.defer(() -> {
          try {
-            snappyRepository.deleteWalletDefaultCardId();
-            cardListStorage.deleteWalletCardList();
+            if(withPaymentCards) {
+               cardListStorage.deleteWalletCardList();
+               snappyRepository.deleteWalletDefaultCardId();
+            }
             snappyRepository.deleteSmartCardFirmware();
             snappyRepository.deleteSmartCardDetails();
             snappyRepository.deleteSmartCard();
